Bo sung thiet ke CSDL lich su giao dich

+++ squirrel-api-NguyenHaNhatMinh-110123264/Models/LichSuGiaoDich.cs
@@ -4,6 +4,6 @@ public class LichSuGiaoDich
 {
     public int Id { get; set; }
     public string SoTaiKhoan { get; set; } = "";
-    public int SoTien { get; set; }
+    public double SoTien { get; set; }
     public string LoaiGiaoDich { get; set; } = "";
 }
